# api/openapi.yaml
openapi: 3.0.3
info:
  title: Network Traffic Visualizer API
  version: 0.1.0
  description: |
    Source of truth for the frontend/API contract (docs/DECISIONS.md D-021).
    TypeScript types are generated from this file (`pnpm gen:api`) and Go
    handlers are validated against it in tests.

    Rates are always `Measurement` objects (D-023). Sampled estimates are
    statistical estimates, never exact throughput.
servers:
  - url: /api/v1
paths:
...
  /globe:
    get:
      operationId: getGlobe
      parameters:
        - { name: grouping, in: query, required: true, schema: { $ref: "#/components/schemas/Grouping" } }
        - { name: source_node_id, in: query, schema: { type: string } }
        - { name: direction, in: query, schema: { $ref: "#/components/schemas/DirectionFilter" } }
        - { name: protocol, in: query, schema: { $ref: "#/components/schemas/Protocol" } }
        - { name: min_bps, in: query, schema: { type: number, minimum: 0 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 1000 } }
        - { $ref: "#/components/parameters/Start" }
        - { $ref: "#/components/parameters/End" }
      responses:
        "200":
          description: External destinations for the current (or requested historical) window.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GlobeResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "501": { $ref: "#/components/responses/HistoryUnavailable" }
        "503": { $ref: "#/components/responses/HistoryError" }
  /globe/destinations/{key}:
    get:
      operationId: getGlobeDestination
      parameters:
        - name: key
          in: path
          required: true
          description: Destination key `<grouping>:<value>` (URL-encoded).
          schema: { type: string }
        - { name: source_node_id, in: query, schema: { type: string } }
        - { name: protocol, in: query, schema: { $ref: "#/components/schemas/Protocol" } }
        - { $ref: "#/components/parameters/Start" }
        - { $ref: "#/components/parameters/End" }
      responses:
        "200":
          description: Destination inspector data.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GlobeDestinationDetail" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "501": { $ref: "#/components/responses/HistoryUnavailable" }
        "503": { $ref: "#/components/responses/HistoryError" }
  /home/traffic:
    get:
      operationId: getHomeTraffic
      parameters:
        - { name: focus_node_id, in: query, schema: { type: string } }
        - { name: destination_key, in: query, schema: { type: string } }
        - { name: vlan_id, in: query, schema: { type: integer, minimum: 0 } }
        - name: device_types
          in: query
          description: Comma-separated device types.
          schema: { type: string }
        - { name: protocol, in: query, schema: { $ref: "#/components/schemas/Protocol" } }
        - { name: min_bps, in: query, schema: { type: number, minimum: 0 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 2000 } }
        - { name: include_inactive, in: query, schema: { type: boolean } }
        - { name: scope, in: query, schema: { $ref: "#/components/schemas/ScopeFilter" } }
        - { $ref: "#/components/parameters/Start" }
        - { $ref: "#/components/parameters/End" }
      responses:
        "200":
          description: Device nodes and logical traffic edges for the current (or requested historical) window.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HomeTrafficResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "501": { $ref: "#/components/responses/HistoryUnavailable" }
        "503": { $ref: "#/components/responses/HistoryError" }
  /home/topology:
    get:
      operationId: getHomeTopology
      parameters:
        - { $ref: "#/components/parameters/Start" }
        - { $ref: "#/components/parameters/End" }
      responses:
        "200":
          description: Known topology with evidence per link.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TopologyResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "501": { $ref: "#/components/responses/HistoryUnavailable" }
        "503": { $ref: "#/components/responses/HistoryError" }
  /devices:
    get:
      operationId: listDevices
      parameters:
        - { name: type, in: query, schema: { $ref: "#/components/schemas/DeviceType" } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/DeviceStatus" } }
        - { name: vlan, in: query, schema: { type: integer, minimum: 0 } }
        - { name: search, in: query, schema: { type: string } }
        - { $ref: "#/components/parameters/Start" }
        - { $ref: "#/components/parameters/End" }
      responses:
        "200":
          description: Known devices.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceListResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "501": { $ref: "#/components/responses/HistoryUnavailable" }
        "503": { $ref: "#/components/responses/HistoryError" }
  /devices/{id}:
    get:
      operationId: getDevice
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: grouping, in: query, schema: { $ref: "#/components/schemas/Grouping" } }
        - { $ref: "#/components/parameters/Start" }
        - { $ref: "#/components/parameters/End" }
      responses:
        "200":
          description: Device inspector data.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DeviceDetail" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "501": { $ref: "#/components/responses/HistoryUnavailable" }
        "503": { $ref: "#/components/responses/HistoryError" }
  /flows:
    get:
      operationId: searchFlows
      description: |
        Investigation endpoint (not the visualization transport). Searches
        the current aggregate window, or [start, end) from history (D-059).
        Results are ordered by estimated bytes; `next_cursor` continues.
      parameters:
        - { name: source, in: query, schema: { type: string } }
        - { name: destination, in: query, schema: { type: string } }
        - { name: source_node_id, in: query, schema: { type: string } }
        - { name: destination_node_id, in: query, schema: { type: string } }
        - { name: protocol, in: query, schema: { $ref: "#/components/schemas/Protocol" } }
        - { name: src_port, in: query, schema: { type: integer, minimum: 0, maximum: 65535 } }
        - { name: dst_port, in: query, schema: { type: integer, minimum: 0, maximum: 65535 } }
        - { name: exporter, in: query, schema: { type: string } }
        - { name: internal_scope, in: query, schema: { $ref: "#/components/schemas/FlowScope" } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 1000 } }
        - { name: cursor, in: query, description: Opaque `next_cursor` of the previous page., schema: { type: string } }
        - { $ref: "#/components/parameters/Start" }
        - { $ref: "#/components/parameters/End" }
      responses:
        "200":
          description: Matching attributed flows.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FlowSearchResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "501": { $ref: "#/components/responses/HistoryUnavailable" }
        "503": { $ref: "#/components/responses/HistoryError" }
  /history/timeline:
    get:
      operationId: getHistoryTimeline
      description: |
        Traffic over time for the timeline (D-059): sampled estimates per
        scope, de-duplicated across observation points, plus boundary counter
        rates where available. Buckets without stored data have has_data=false.
      parameters:
        - { name: start, in: query, required: true, schema: { $ref: "#/components/schemas/Timestamp" } }
        - { name: end, in: query, required: true, schema: { $ref: "#/components/schemas/Timestamp" } }
        - { name: max_points, in: query, schema: { type: integer, minimum: 2, maximum: 720 } }
      responses:
        "200":
          description: Timeline buckets.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HistoryTimelineResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "501": { $ref: "#/components/responses/HistoryUnavailable" }
        "503": { $ref: "#/components/responses/HistoryError" }
...
components:
  parameters:
    Start:
      name: start
      in: query
      description: Start of a historical window (RFC3339, inclusive). Requires `end`; omit both for the live window.
      schema: { $ref: "#/components/schemas/Timestamp" }
    End:
      name: end
      in: query
      description: End of a historical window (RFC3339, exclusive). The range is aligned outwards to the stored resolution.
      schema: { $ref: "#/components/schemas/Timestamp" }
  responses:
    HistoryUnavailable:
      description: History is not enabled on this server.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    HistoryError:
      description: The history store could not be queried.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    BadRequest:
      description: Invalid parameter.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
    NotFound:
      description: Not found in the current window.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
  schemas:
...
    Timestamp:
      type: string
      format: date-time
      description: RFC3339 UTC.
...
    TimeWindow:
      type: object
      required: [start, end]
      properties:
        start: { $ref: "#/components/schemas/Timestamp" }
        end: { $ref: "#/components/schemas/Timestamp" }
...
    DirectionFilter:
      type: string
      enum: [inbound, outbound, both]
    Protocol:
      type: string
      enum: [tcp, udp, icmp, other]
    ProtocolShare:
      type: object
      required: [protocol, port, service, bps]
      properties:
        protocol: { $ref: "#/components/schemas/Protocol" }
        port: { type: integer, nullable: true }
        service: { type: string, nullable: true }
        bps: { $ref: "#/components/schemas/Measurement" }
    ObservationPoint:
      type: object
      required: [exporter_id, exporter_name, input_if_index, output_if_index, used_for_aggregate, sampling_rate]
      properties:
        exporter_id: { type: string }
        exporter_name: { type: string }
        input_if_index: { type: integer, nullable: true }
        output_if_index: { type: integer, nullable: true }
        used_for_aggregate:
          type: boolean
          description: True for the observation point whose values were used (others are not summed, D-024).
        sampling_rate: { type: integer }
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }

    # ------------------------------------------------------------- status
...
    HistoryTimelinePoint:
      type: object
      required: [start, has_data, inbound, outbound, internal, wan_download, wan_upload]
      properties:
        start: { $ref: "#/components/schemas/Timestamp" }
        has_data: { type: boolean }
        inbound: { nullable: true, allOf: [{ $ref: "#/components/schemas/Measurement" }], description: "External → internal (download), sampled estimate." }
        outbound: { nullable: true, allOf: [{ $ref: "#/components/schemas/Measurement" }], description: "Internal → external (upload), sampled estimate." }
        internal: { nullable: true, allOf: [{ $ref: "#/components/schemas/Measurement" }], description: "Internal ↔ internal, sampled estimate." }
        wan_download: { nullable: true, allOf: [{ $ref: "#/components/schemas/Measurement" }], description: Boundary counter rate. }
        wan_upload: { nullable: true, allOf: [{ $ref: "#/components/schemas/Measurement" }], description: Boundary counter rate. }
    HistoryTimelineResponse:
      type: object
      required: [window, step_seconds, points]
      properties:
        window: { $ref: "#/components/schemas/TimeWindow" }
        step_seconds: { type: integer }
        points: { type: array, items: { $ref: "#/components/schemas/HistoryTimelinePoint" } }

    # -------------------------------------------------------------- globe
    Grouping:
      type: string
      enum: [country, city, asn, ip]
    GlobeOrigin:
      type: object
      required: [label, latitude, longitude, precision]
      properties:
        label: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        precision: { type: string, enum: [country, region, city, custom] }
...
    GlobeDestination:
      type: object
      required:
        [key, grouping, label, ip, country_code, country_name, city, location, asn, organization,
         inbound_bps, outbound_bps, source_device_count, source_node_ids, last_seen]
      properties:
        key: { type: string, description: "`<grouping>:<value>`, parse by first colon." }
        grouping: { $ref: "#/components/schemas/Grouping" }
        label: { type: string }
        ip: { type: string, nullable: true }
        country_code: { type: string, nullable: true }
        country_name: { type: string, nullable: true }
        city: { type: string, nullable: true }
        location:
          nullable: true
          description: null = unknown; never substituted with a placeholder.
          allOf: [{ $ref: "#/components/schemas/GeoLocation" }]
        asn: { type: integer, nullable: true }
        organization: { type: string, nullable: true }
        inbound_bps: { $ref: "#/components/schemas/Measurement" }
        outbound_bps: { $ref: "#/components/schemas/Measurement" }
        source_device_count: { type: integer }
        source_node_ids: { type: array, items: { type: string } }
        last_seen: { $ref: "#/components/schemas/Timestamp" }
    WanSummary:
      type: object
      required: [download, upload, basis]
      properties:
        download: { $ref: "#/components/schemas/Measurement" }
        upload: { $ref: "#/components/schemas/Measurement" }
        basis: { type: string, enum: [boundary_counter, sampled_sum] }
    GlobeResponse:
      type: object
      required: [window, server_time, origin, grouping, destinations, truncated_count, summary]
      properties:
        window: { $ref: "#/components/schemas/TimeWindow" }
        server_time: { $ref: "#/components/schemas/Timestamp" }
        origin: { $ref: "#/components/schemas/GlobeOrigin" }
        grouping: { $ref: "#/components/schemas/Grouping" }
        destinations: { type: array, items: { $ref: "#/components/schemas/GlobeDestination" } }
        truncated_count: { type: integer }
        summary: { $ref: "#/components/schemas/WanSummary" }
    DestinationSource:
      type: object
      required: [node_id, label, device_type, resolved, inbound_bps, outbound_bps]
      properties:
        node_id: { type: string }
        label: { type: string }
        device_type: { $ref: "#/components/schemas/DeviceType" }
        resolved: { type: boolean }
        inbound_bps: { $ref: "#/components/schemas/Measurement" }
        outbound_bps: { $ref: "#/components/schemas/Measurement" }
    DestinationMember:
      type: object
      required: [ip, city, country_code, total_bps]
      properties:
        ip: { type: string }
        city: { type: string, nullable: true }
        country_code: { type: string, nullable: true }
        total_bps: { $ref: "#/components/schemas/Measurement" }
    GlobeDestinationDetail:
      type: object
      required: [destination, top_sources, top_protocols, members, observation_points]
      properties:
        destination: { $ref: "#/components/schemas/GlobeDestination" }
        top_sources: { type: array, items: { $ref: "#/components/schemas/DestinationSource" } }
        top_protocols: { type: array, items: { $ref: "#/components/schemas/ProtocolShare" } }
        members: { type: array, items: { $ref: "#/components/schemas/DestinationMember" } }
        observation_points: { type: array, items: { $ref: "#/components/schemas/ObservationPoint" } }

    # --------------------------------------------------------------- home
    DeviceType:
      type: string
      enum: [internet, router, firewall, switch, wireless_ap, server, pc, smartphone, iot, vm, kubernetes_node, unknown]
    DeviceStatus:
      type: string
      enum: [online, stale, offline, unknown]
    HomeNodeKind:
      type: string
      enum: [device, unresolved_endpoint, internet, external_destination]
    HomeNode:
      type: object
      required: [id, kind, label, device_type, status, addresses, vlan_id, network_name, rx_bps, tx_bps, last_seen]
      properties:
        id: { type: string, description: "Device id, `ep:<ip>`, `internet` or `dst:<key>` (D-027)." }
        kind: { $ref: "#/components/schemas/HomeNodeKind" }
        label: { type: string }
        device_type: { $ref: "#/components/schemas/DeviceType" }
        status: { $ref: "#/components/schemas/DeviceStatus" }
        addresses: { type: array, items: { type: string } }
        vlan_id: { type: integer, nullable: true }
        network_name: { type: string, nullable: true }
        rx_bps:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/Measurement" }]
        tx_bps:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/Measurement" }]
        last_seen: { type: string, format: date-time, nullable: true }
...
    ScopeFilter:
      type: string
      enum: [internal, external, both]
    TrafficEdge:
      type: object
      description: Logical communication between two endpoints, not a physical path (D-032).
      required: [id, source, target, scope, forward_bps, reverse_bps, top_protocols, observation_points, last_seen]
      properties:
        id: { type: string }
        source: { type: string }
        target: { type: string }
        scope: { $ref: "#/components/schemas/TrafficScope" }
        forward_bps: { $ref: "#/components/schemas/Measurement" }
        reverse_bps: { $ref: "#/components/schemas/Measurement" }
        top_protocols: { type: array, items: { $ref: "#/components/schemas/ProtocolShare" } }
        observation_points: { type: array, items: { $ref: "#/components/schemas/ObservationPoint" } }
        last_seen: { $ref: "#/components/schemas/Timestamp" }
    HomeSummary:
      type: object
      required: [wan_total, internal_estimated, device_count, online_count]
      properties:
        wan_total: { $ref: "#/components/schemas/Measurement" }
        internal_estimated: { $ref: "#/components/schemas/Measurement" }
        device_count: { type: integer }
        online_count: { type: integer }
    DestinationContext:
      type: object
      required: [key, label, source_node_ids]
      properties:
        key: { type: string }
        label: { type: string }
        source_node_ids: { type: array, items: { type: string } }
    HomeTrafficResponse:
      type: object
      required: [window, server_time, nodes, edges, truncated_count, destination_context, summary]
      properties:
        window: { $ref: "#/components/schemas/TimeWindow" }
        server_time: { $ref: "#/components/schemas/Timestamp" }
        nodes: { type: array, items: { $ref: "#/components/schemas/HomeNode" } }
        edges: { type: array, items: { $ref: "#/components/schemas/TrafficEdge" } }
        truncated_count: { type: integer }
        destination_context:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/DestinationContext" }]
        summary: { $ref: "#/components/schemas/HomeSummary" }
...
    TopologyLink:
      type: object
      required: [id, source_node_id, target_node_id, link_type, evidence, confidence, source_interface, target_interface]
      properties:
        id: { type: string }
        source_node_id: { type: string }
        target_node_id: { type: string }
        link_type: { $ref: "#/components/schemas/TopologyLinkType" }
        evidence: { $ref: "#/components/schemas/TopologyEvidence" }
        confidence: { type: number, minimum: 0, maximum: 1 }
        source_interface: { type: string, nullable: true }
        target_interface: { type: string, nullable: true }
    TopologyResponse:
      type: object
      required: [nodes, links, generated_at]
      properties:
        nodes: { type: array, items: { $ref: "#/components/schemas/HomeNode" } }
        links: { type: array, items: { $ref: "#/components/schemas/TopologyLink" } }
        generated_at: { $ref: "#/components/schemas/Timestamp" }

    # ------------------------------------------------------------ devices
    DeviceAddress:
      type: object
      required: [address, family, source]
      properties:
        address: { type: string }
        family: { $ref: "#/components/schemas/AddressFamily" }
        source: { type: string, enum: [static, dhcp, arp, nd, dns, wlc, flow, manual] }
    DevicePeer:
      type: object
      required: [node_id, label, device_type, tx_bps, rx_bps]
      properties:
        node_id: { type: string }
        label: { type: string }
        device_type: { $ref: "#/components/schemas/DeviceType" }
        tx_bps: { $ref: "#/components/schemas/Measurement" }
        rx_bps: { $ref: "#/components/schemas/Measurement" }
    DeviceExternalDestination:
      type: object
      required: [key, label, country_code, outbound_bps, inbound_bps]
      properties:
        key: { type: string }
        label: { type: string }
        country_code: { type: string, nullable: true }
        outbound_bps: { $ref: "#/components/schemas/Measurement" }
        inbound_bps: { $ref: "#/components/schemas/Measurement" }
    DeviceAttachment:
      type: object
      required: [via_node_id, via_label, interface, evidence]
      properties:
        via_node_id: { type: string }
        via_label: { type: string }
        interface: { type: string, nullable: true }
        evidence: { $ref: "#/components/schemas/TopologyEvidence" }
    DeviceDetail:
      type: object
      required: [node, addresses, macs, ssid, attachment, vendor, model, top_internal_peers, top_external_destinations]
      properties:
        node: { $ref: "#/components/schemas/HomeNode" }
        addresses: { type: array, items: { $ref: "#/components/schemas/DeviceAddress" } }
        macs: { type: array, items: { type: string } }
        ssid: { type: string, nullable: true }
        attachment:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/DeviceAttachment" }]
        vendor: { type: string, nullable: true }
        model: { type: string, nullable: true }
        top_internal_peers: { type: array, items: { $ref: "#/components/schemas/DevicePeer" } }
        top_external_destinations: { type: array, items: { $ref: "#/components/schemas/DeviceExternalDestination" } }
    DeviceListResponse:
      type: object
      required: [devices]
      properties:
        devices: { type: array, items: { $ref: "#/components/schemas/HomeNode" } }

    # -------------------------------------------------------------- flows
    FlowScope:
      type: string
      enum: [internal, external, transit]
    FlowRecord:
      type: object
      required:
        [src_ip, dst_ip, protocol, src_port, dst_port, scope, direction, src_node_id, dst_node_id,
         bps, exporter_id, sampling_rate, last_seen]
      properties:
        src_ip: { type: string }
        dst_ip: { type: string }
        protocol: { $ref: "#/components/schemas/Protocol" }
        src_port: { type: integer, nullable: true }
        dst_port: { type: integer, nullable: true }
        scope: { $ref: "#/components/schemas/FlowScope" }
        direction:
          nullable: true
          allOf: [{ $ref: "#/components/schemas/Direction" }]
        src_node_id: { type: string, nullable: true }
        dst_node_id: { type: string, nullable: true }
        bps: { $ref: "#/components/schemas/Measurement" }
        exporter_id: { type: string, description: Observation point used for the value. }
        sampling_rate: { type: integer }
        last_seen: { type: string, format: date-time, nullable: true }
    FlowSearchResponse:
      type: object
      required: [window, flows, truncated_count, next_cursor]
      properties:
        window: { $ref: "#/components/schemas/TimeWindow" }
        flows: { type: array, items: { $ref: "#/components/schemas/FlowRecord" } }
        truncated_count: { type: integer, description: Matches after this page. }
        next_cursor: { type: string, nullable: true, description: Pass as `cursor` for the next page; null on the last page. }

    # ----------------------------------------------------------- realtime
...